Switch to generic lazy globals
