+++ packages/signals/src/global/mod.rs
@@ -2,7 +2,8 @@ use dioxus_core::{
     prelude::{provide_root_context, try_consume_context},
     ScopeId,
 };
-use std::{any::Any, cell::RefCell, collections::HashMap, panic::Location, rc::Rc};
+use generational_box::BorrowResult;
+use std::{any::Any, cell::RefCell, collections::HashMap, ops::Deref, panic::Location, rc::Rc};
 
 mod memo;
 pub use memo::*;
@@ -10,7 +11,7 @@ pub use memo::*;
 mod signal;
 pub use signal::*;
 
-use crate::Signal;
+use crate::{Readable, ReadableRef, Signal, Writable, WritableRef};
 
 /// A trait for an item that can be constructed from an initialization function
 pub trait InitializeFromFunction<T> {
@@ -29,6 +30,95 @@ pub struct LazyGlobal<T, R = T> {
     constructor: fn() -> R,
     key: GlobalKey<'static>,
     phantom: std::marker::PhantomData<fn() -> T>,
+}
+
+/// Allow calling a signal with signal() syntax
+///
+/// Currently only limited to copy types, though could probably specialize for string/arc/rc
+impl<T: Clone + 'static, R: Clone + 'static> Deref for LazyGlobal<T, R>
+where
+    T: Readable<Target = R> + InitializeFromFunction<R>,
+{
+    type Target = dyn Fn() -> R;
+
+    fn deref(&self) -> &Self::Target {
+        Readable::deref_impl(&self.resolve())
+    }
+}
+
+impl<T: Clone + 'static, R: 'static> Readable for LazyGlobal<T, R>
+where
+    T: Readable<Target = R> + InitializeFromFunction<R>,
+{
+    type Target = R;
+    type Storage = T::Storage;
+
+    #[track_caller]
+    fn try_read_unchecked(
+        &self,
+    ) -> Result<ReadableRef<'static, Self>, generational_box::BorrowError> {
+        self.resolve().try_read_unchecked()
+    }
+
+    #[track_caller]
+    fn try_peek_unchecked(&self) -> BorrowResult<ReadableRef<'static, Self>> {
+        self.resolve().try_peek_unchecked()
+    }
+}
+
+impl<T: Clone + 'static, R: 'static> Writable for LazyGlobal<T, R>
+where
+    T: Writable<Target = R> + InitializeFromFunction<R>,
+{
+    type Mut<'a, Read: ?Sized + 'static> = T::Mut<'a, Read>;
+
+    fn map_mut<I: ?Sized, U: ?Sized + 'static, F: FnOnce(&mut I) -> &mut U>(
+        ref_: Self::Mut<'_, I>,
+        f: F,
+    ) -> Self::Mut<'_, U> {
+        T::map_mut(ref_, f)
+    }
+
+    fn try_map_mut<
+        I: ?Sized + 'static,
+        U: ?Sized + 'static,
+        F: FnOnce(&mut I) -> Option<&mut U>,
+    >(
+        ref_: Self::Mut<'_, I>,
+        f: F,
+    ) -> Option<Self::Mut<'_, U>> {
+        T::try_map_mut(ref_, f)
+    }
+
+    fn downcast_lifetime_mut<'a: 'b, 'b, Read: ?Sized + 'static>(
+        mut_: Self::Mut<'a, Read>,
+    ) -> Self::Mut<'b, Read> {
+        T::downcast_lifetime_mut(mut_)
+    }
+
+    #[track_caller]
+    fn try_write_unchecked(
+        &self,
+    ) -> Result<WritableRef<'static, Self>, generational_box::BorrowMutError> {
+        self.resolve().try_write_unchecked()
+    }
+}
+
+impl<T: Clone + 'static, R: 'static> LazyGlobal<T, R>
+where
+    T: Writable<Target = R> + InitializeFromFunction<R>,
+{
+    /// Write this value
+    pub fn write(&self) -> T::Mut<'static, R> {
+        self.resolve().try_write_unchecked().unwrap()
+    }
+
+    /// Run a closure with a mutable reference to the signal's value.
+    /// If the signal has been dropped, this will panic.
+    #[track_caller]
+    pub fn with_mut<O>(&self, f: impl FnOnce(&mut R) -> O) -> O {
+        self.resolve().with_mut(f)
+    }
 }
 
 impl<T: Clone + 'static, R> LazyGlobal<T, R>
